v0.2.3: Quick Add NCC/SP, Draft Persistence, Product Sorting, and Corrected Return Accounting in Reports

+++ backend/app/schemas/batch.py
@@ -12,5 +12,5 @@ class BatchOut(BaseModel):
     batch_code: str
     expiry_date: date
     import_price: Decimal
-    quantity_remaining: float
+    quantity_remaining: int
     created_at: datetime
